chore: Move 'Container' and 'Layout' to components
