+++ source/components/Container/Container.tsx
@@ -4,10 +4,8 @@ interface IContainer {
   children?: ReactNode;
 }
 
-const Container: FC<IContainer> = ({ children }) => (
+export const Container: FC<IContainer> = ({ children }) => (
   <div className="m-0 mx-auto p-0 w-full min-w-popup h-full min-h-popup font-poppins text-xl md:max-w-2xl">
     {children}
   </div>
 );
-
-export default Container;
