+++ source/components/Layout/AuthViewLayout.tsx
@@ -2,7 +2,7 @@ import React, { FC } from 'react';
 import { Icon, IconButton, Tooltip } from 'components/index';
 import { useUtils, useBrowser } from 'hooks/index';
 
-import { Header } from '../Header';
+import { Header } from '../../pages/Header';
 
 interface IAuthViewLayout {
   background?: string;
